Moved byzantine_client_fed_avg.py and byzantine_weight_client_fed_avg.py to base_client_fed_avg
